Fix loading crates with a hyphen when using `extern crate`
As of RFC0940[1], crates shouldn't contain any hyphens anymore. This
will cause a syntax error in Rust, however `cargo` is supposed to
internally replace hyphens with underscores.

While this is not a problem when using the `:dep`-functionality, this
will fail when using the `extern crate`-hook of the REPL since it fails
with a syntax error when using a dash, but can't find the crate if it's
specified with an underscore.

Unfortunately, the `Cargo.toml` seems to be sensitve about hyphen vs.
underscore (although even the crates.io API isn't), so we need to check
for the actual crate's name first to make sure that e.g. `rand_core`
isn't replaced by `rand-core`.

Resolves #61

[1] https://rust-lang.github.io/rfcs/0940-hyphens-considered-harmful.html