Always show full original lines when reporting errors
Before, if an input line had multiple statements, we would report errors
showing just the statement that contained the error. Now we show the
full line.

Primary motivation is that it simplifies the code, since we no longer
need to track spans with two different reference points.

Also improved error span handling in the presence of graphemes in the REPL.