Rewrite Cargo.toml after we add tokio as a dependency
Normally we add dependencies before we write Cargo.toml, however adding
tokio in response to async code is a special case as it happens after
we've tried to compile the code, so we've already written Cargo.toml.
There are probably several alternative ways we could have fixed this.
One way would be to have dependency validation rename Cargo.toml before
overwriting it, then move it back when it's done. But that would be more
code.

Fixes #172