Change preserve_vars_on_panic default back to true
It does cause a slight increase in compilation time, probably due to
rustc needing to pull in more of the standard library. However, this is
really needed to avoid trying to propagate the panic past an FFI
boundary, which is undefined behavior. See #226

With the previous commit, preserving variables is now much more
consistent and works much better than it did when it was turned off by
default.