Delete code that tries to use compiler errors to determine types
The last thing that this code was needed for was arrays, but
rust-analyzer supports determining the types of arrays now.

The remaining cases where rust-analyzer can't determine the type, the
compiler gives us types that aren't fully qualified, which doesn't help
us and just gives us confusing error messages.

So now, if rust-analyzer fails to determine a variable type, we bail and
ask the user to add an explicit type annotation.