Make building no-std crates as dylibs more robust
Before we were relying on enabling the 'std' feature of crates. Instead,
we now just tell rustc that the 'core' crate can be found in libstd.so
(or equivalent). Since everything depends on the 'core' crate and the
'std' crate is a superset of the 'core' crate, this should hopefully
have no unintended side effects.

Issue #331