jupyter: Ensure iopub idle isn't sent before stdout output
Because we use stdout to notify completion of the user code, we can
ensure that any output printed to stdout prior to the user code
completing is sent to the iopub before we send the idle message.

We can't really do a similar thing with stderr because there's no
control message sent to stderr to indicate the end of execution. However
doing this at least for stdout is better than nothing.

Issue #353